Remove pointless casts

+++ bitset.go
@@ -36,7 +36,7 @@ func (b Bitset) Subset(start, end uint) Bitset {
 		panic(errEndLessThanStart)
 	}
 
-	if end > uint(b.Len()) {
+	if end > b.Len() {
 		panic(errOutOfRange)
 	}
